requirements.txt
allure-pytest
pytest
webdriver-manager
selenium
pyautogui